refactor(frontend): decompose ActiveWatchlist.tsx into ActiveWatchlist/ directory
Extracted useWatchlistLive hook (WS types + reconnect logic) to hooks/.
Split page into AlertBadges (severity badge), WatchlistTable (PriceCell,
SessionCell, WatchlistRow with inline mutations, table), and shell index.tsx
(AddSymbolForm, header, limit warning, card layout). All files under 300 LOC.
